requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.0.1
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
python-multipart==0.0.6
aiofiles==23.2.1
locust==2.17.0
jupyter==1.0.0
notebook==7.0.6
python-dotenv==1.0.0
